We're going to build a very simple todo list using React. You'll see how easy React is to get started with.
To start, I've just included a `lib` directory with a couple scripts we'll include. One is the React library itself. The other is a very simple shim called `JSM` that let's us write clean templates without having to learn `JSX`, or to worry about the complexities of building our project whenever there are changes. We're keeping it simple, remember?