Now we connect up the simplest possible React application to make sure everything is working properly.
We create a React component called `App`, with a `render` method. This method returns the markup for the `App` component.

See how we're using `jsm` to define the markup? It's pretty simple once you get used to it. You write nested arrays, where the first element is the tag name, the second is an object of attributes, and the third is the content.

Finally, we tell React to render our `App` component into the element with an id of `app`. Of course, we had to add that element to the HTML page so there was something to render to.