Render a todo list
Next let's render a static list of todos. We store our todos as an array in the components state. Create some initial todos by defining the `getInitialState` method.

In the `render` method, we're listing those todos as list items. We create an `items` array of list items for is todo, by mapping over the todos. And, we add a `ul` to our `div` that includes those items.