Adding Todos
Of course, we want to add our own todos to our list.

Add a text input to our component. Notice that we add a `ref` attribute to the text input. This makes it easy to reference that element later when we want its value.

Now add a button. Notice the `onClick` attribute. React will call the function we pass to it whenever the button is clicked.

Finally, we need to define the `addTodo` function that the button will trigger. We want to do several things here: add a new todo to the list, clear the input, and re-focus the input. We get a reference to the input element using `React.findDOMNode`, together with the `ref` that we defined earlier.

Try it out. If you reload `index.html`, you should be able to add todo items now.