Removing todos
How about removing todos from our list when they are complete?

Add a link to the list item for each todo. There are a couple of tricks here. First, we bound the value of `this` in the `map` function to the component, so that we can use it inside the callback function. Second, we used `bind` to make sure the index is passed to our `removeTodo` method when it is called by React. This is important so that we know *which* todo was clicked.

The `removeTodo` method is quite simple. We use `splice` to remove the todo with the given index.